fix: Remove 'clear' button when value is empty string
